requirements.txt
blinker==1.4
bson==0.4.3
click==6.6
dominate==2.3.1
Flask==0.12
Flask-Bootstrap==3.3.7.0
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-WTF==0.13.1
itsdangerous==0.24
Jinja2==2.8
jpush==3.2.1
MarkupSafe==0.23
MySQL-python==1.2.5
Pillow==3.4.2
pymongo==3.4.0
pytz==2016.10
requests==2.12.4
six==1.10.0
SQLAlchemy==1.1.4
visitor==0.1.3
Werkzeug==0.11.13
WTForms==2.1
xlwt==1.2.0
